requirements.txt
Django==1.11.12
django-admin-interface==0.6.3
django-ckeditor==5.4.0
django-colorfield==0.1.14
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.0.0
django-versatileimagefield==1.9
Pillow==5.0.0
pytz==2018.4
